requirements.txt
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cors==1.0.1
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SocketIO==5.3.7
future==1.0.0
gevent==24.2.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
netifaces==0.10.6
PySocks==1.7.1
python-engineio==4.9.1
python-socketio==5.11.4
requests==2.32.3
requests-file==2.1.0
setuptools==3.3
simple-websocket==1.0.0
socketio==0.2.1
tldextract==5.1.2
urllib3==2.2.3
Werkzeug==3.0.4
wsproto==1.2.0
zope.event==5.0
zope.interface==7.0.3
